requirements.txt
streamlit
bcrypt
scikit-learn
joblib
folium
streamlit-folium
geopy
python-dotenv
hypothesis
keyboard
